Move contructing the association table to separate file
